Use class based views to make them more easily extensible
This commit updates the views related to device auth flow from
functions to class-based views. This makes it easy to import
them in other projects and only overwrite small bits of functionality
(specifically looking at templates and context_data).

A 3rd view is added for the final step where the user is presented
with the status of the device after they approve or deny.

The views now also have a more "standard" django form behaviour, changes
being relfected in the tests.